Adapt sja1000_raspi to use device tree
The driver for our custom CAN extension board for the Raspberry Pi needs
to know the location of the GPIO registers in memory.

The previous version simply used the constant GPIO_BASE defined in the
appropriate mach/platform.h file for the Raspberry Pi. This file has
been removed since because the developers want to enforce the use of
Device Tree for hadware configuration instead of hardcoded addresses
that require a kernel patch every time a new board is released.

Therefore, this module has been adapted to query the memory mapping of
the GPIOs from the device tree using its API.

This implementation differs from the standard "device tree tutorial
solution" because the CAN extension board is not represented in the
device tree. Usually the tutorials use the struct platform_device* given
to a driver's xxx_probe() method to retrieve the memory information, but
in this case the memory information there is empty because the kernel
does not "know" that sja1000_raspi needs to access the GPIOs.

To mitigate this, the probe method uses of_find_node_by_name() to get
the device tree node for the GPIO and then of_address_to_resource() to
query the memory location in a standard kernel struct resource.

Directly reading the memory location from the device tree is possible,
but does not work because there are different views on the address
space: The device tree specifies a bus address as it is seen by the
memory controller. As the memory is split between the GPU and the CPU,
the same memory location has another address as kernel physical address.
The function of_address_to_resource() is aware of this and does the
conversion from bus address to physical address automatically. In the
end, ioremap() is called with the physical address to map the GPIO
registers into the kernel virtual memory and make the necessary
amendments to the kernel page table, so finally there is a register base
address that can be given to the main sja1000 driver.

The implementation has been tested successfully on a Raspberry Pi 2
Model B.